+++ modules/App/layouts/app.php
@@ -99,7 +99,7 @@
         </aside>
         <main class="kiss-flex-1">
 
-            <app-header>
+            <app-header data-sticky="true">
                 <kiss-container class="kiss-flex kiss-flex-middle">
                     <a class="kiss-display-block kiss-margin-small-right kiss-visible@m" href="<?=$this->route('/')?>">
                         <img class="app-logo kiss-margin-auto" src="<?=$this->helper('theme')->logo()?>" alt="Logo" style="height:30px;width:auto;">