+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fix broken assets loading for asset field with multiple values
 - Fix referenced assets updates for content tree items
 - Fix mime type of generated thumbnails in assets manager
+- Add fields outline for better navigation
 
 ## 2.7.0 (2023-09-15)